Fix build attempt 1
conda was updating python to 3.8 on the pandas install. That shouldn't
cause problems but here we are. In kb-sdk test nose and jinja2 couldn't be found
after the conda install. pip seems to work.

Removed a path statement that doesn't appear to be necessary and was
making pip install pandas in the wrong place.